requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django==2.2.3
django-rest-knox==4.1.0
djangorestframework==3.12.1
pycparser==2.20
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
